Extract Workspace Dock (eval/REPL/Transcript/Changes/git) out of WorkspaceLive (BT-3295)
Moves the eval/Workspace tab, REPL tab, and Changes/Git tab handle_event and
handle_async(:git_load, …) clauses out of WorkspaceLive into a new
BtAttachWeb.Live.Dock module, following the BT-3291 Inspector extraction
pattern: WorkspaceLive delegates by event name (@dock_events) and forwards
handle_async(:git_load, …) unchanged, and the render template still lives in
WorkspaceLive, calling Dock's few template-facing helpers (ws_selection?/1,
workspace_editor_id/0, repl_preview/1, repl_input_id/0) qualified.

Cross-cutting helpers the moved handlers still need from panes not yet
extracted (Method Editor tabs, System Browser, Tests pane) were made public
on WorkspaceLive rather than duplicated: assign_changes/1, status_error/2,
maybe_refresh_git/1, clamp_offset/1, symbol_rows/1, open_class/2,
discover_test_classes/2, refresh_after_source_change/1,
reload_reverted_def_buffers/2, and a new valid_class_name?/1.

Adds direct BtAttachWeb.Live.DockTest unit tests (StubWorkspaceClient, no
live workspace node) covering eval action routing, the REPL history ring,
git stage/unstage/commit/revert error paths, and the destructive-flush
confirmation path — branches previously reachable only through the
:workspace-tagged WorkspaceLiveTest.
